feat(area): extract area objects to area model
