+++ antarest/study/business/table_mode_management.py
@@ -20,13 +20,14 @@ from typing_extensions import override
 
 from antarest.core.exceptions import ChildNotFoundError
 from antarest.core.model import JSON
-from antarest.study.business.area_management import AreaManager, AreaOutput
+from antarest.study.business.area_management import AreaManager
 from antarest.study.business.areas.renewable_management import RenewableClusterInput, RenewableManager
 from antarest.study.business.areas.st_storage_management import STStorageInput, STStorageManager
 from antarest.study.business.areas.thermal_management import ThermalClusterInput, ThermalManager
 from antarest.study.business.binding_constraint_management import BindingConstraintManager, ConstraintInput
 from antarest.study.business.enum_ignore_case import EnumIgnoreCase
 from antarest.study.business.link_management import LinkManager
+from antarest.study.business.model.area_model import AreaOutput
 from antarest.study.business.model.link_model import LinkBaseDTO
 from antarest.study.model import STUDY_VERSION_8_2, RawStudy
